+++ src/lexer/utils/index.ts
@@ -4,3 +4,4 @@ export * from "./word";
 export * as javascript from "./javascript";
 export * as json from "./json";
 export * from "./numeric";
+export * from "./subaction";